Added testing script
Added testing script and remove unnecessary files.
