--- brownie/BEN-Pet.json
@@ -1,1 +1,0 @@
-{"anonymous":false,"inputs":[{"indexed":false,"internalType":"address","name":"NBT","type":"address"},{"indexed":false,"internalType":"string","name":"prompt","type":"string"},{"indexed":false,"internalType":"uint256","name":"tokenId","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"value","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"scratchId","type":"uint256"}],"name":"Pet","type":"event"}
