--- brownie/CultureCoin-Pay.json
@@ -1,1 +1,0 @@
-{"anonymous":false,"inputs":[{"indexed":false,"internalType":"address","name":"who","type":"address"},{"indexed":false,"internalType":"uint256","name":"amount","type":"uint256"}],"name":"Pay","type":"event"}
