--- brownie/GamblersUnionBEN-ContestStarted.json
@@ -1,1 +1,0 @@
-{"anonymous":false,"inputs":[{"indexed":false,"internalType":"uint256","name":"contestId","type":"uint256"},{"indexed":false,"internalType":"address","name":"NBT","type":"address"},{"indexed":false,"internalType":"uint256","name":"closeTime","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"rightTokenId","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"leftTokenId","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"rightStart","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"leftStart","type":"uint256"}],"name":"ContestStarted","type":"event"}
