--- brownie/MarketPlace-OfferingPlaced.json
@@ -1,1 +1,0 @@
-{"anonymous":false,"inputs":[{"indexed":true,"internalType":"bytes32","name":"offeringId","type":"bytes32"},{"indexed":true,"internalType":"address","name":"hostContract","type":"address"},{"indexed":true,"internalType":"address","name":"offerer","type":"address"},{"indexed":false,"internalType":"uint256","name":"tokenId","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"price","type":"uint256"},{"indexed":false,"internalType":"string","name":"uri","type":"string"}],"name":"OfferingPlaced","type":"event"}
