--- brownie/PrintingPress-BookContract.json
@@ -1,1 +1,0 @@
-{"anonymous":false,"inputs":[{"indexed":false,"internalType":"address","name":"who","type":"address"},{"indexed":false,"internalType":"address","name":"what","type":"address"}],"name":"BookContract","type":"event"}
