--- brownie/key.json
@@ -1,1 +1,0 @@
-{"address": "183a3e96a8d52e4f4b07688acfa0fcf50a4cff02", "crypto": {"cipher": "aes-128-ctr", "cipherparams": {"iv": "5b1aa6a0025a9fc97b1d7f41f6b1ab25"}, "ciphertext": "3fd1994f6f9896ff200e31112c26f3de1916c8b4cb1b5e10f6091f14aea979d0", "kdf": "scrypt", "kdfparams": {"dklen": 32, "n": 262144, "r": 1, "p": 8, "salt": "ef5c2482430a0e178eea8cdb3817cd41"}, "mac": "bb9afdad728e87d8b79c61edeea6584f2524aa5844edab550094a8ec19bf4a24"}, "id": "ef36a108-092f-4e8d-8c75-84e5188be437", "version": 3}
