+++ brownie/tests/greatLibrary_test.py
@@ -12,6 +12,177 @@ deployAmount = 2 * 210100027 * OneCC
 from scripts.helpful_scripts import encode_function_data
 
 @pytest.fixture(scope="session")
+def HEROS2(CC, PP, MP, BookTradable, Hero, spells, loot, items):
+    LOG.info("test_hero")
+
+
+    # address _NBT, string memory _prompt, uint256 _tokenId
+    tx = PP.newBookContract("exampleNBT", "BTtestNBT", MP.address, "urlgoeshere", True, maxint, OneCC, 1, accounts[0].address)
+    exampleBook = BookTradable.at(tx.events[2]["what"])
+    tx = PP.newBookContract("exampleNBT", "BMtestNBT", MP.address, "urlgoeshere", True, maxint, OneCC, 1, accounts[0].address)
+    exampleBookMark = BookTradable.at(tx.events[2]["what"])
+
+    exampleBookMark.setAddon(PP.address, True, {'from': accounts[0].address})
+    exampleBook.setAddon(PP.address, True, {'from': accounts[0].address})
+
+    exampleBookMark.setRewardContract(exampleBook.address, {'from': accounts[0].address})
+    exampleBook.setAddon(exampleBookMark.address, True, {'from': accounts[0].address})
+
+    mintTx = PP.delegateMinter(accounts[0].address, exampleBookMark.address, 5, 0, 0, {'from': accounts[0].address})
+    LOG.info("Mint tx: " + str(mintTx))
+
+    buyTx = PP.buyBook(exampleBook.address, {'from': accounts[0].address, 'value': OneCC})
+    LOG.info("Buy tx: " + str(buyTx))
+
+    buyBMTx = PP.buyBook(exampleBookMark.address, {'from': accounts[0].address, 'value': OneCC})
+    LOG.info("Buy tx: " + str(buyBMTx))
+
+    #address _NBT, string memory _prompt, uint256 _tokenId This the hardbound?
+    tx = PP.newBookContract("exampleNBT", "BTtestNBT", MP.address, "urlgoeshere", True, maxint, OneCC, 1, accounts[0].address)
+
+    ##address _cCA, address _cultureCoin, address _nbt, address _registryAddress
+    hero = accounts[0].deploy(Hero, accounts[0].address, CC.address, exampleBookMark.address, MP.address, spells.address)
+    LOG.info("The heros are here: " + str(hero))
+
+    CC.setAddon(hero.address, True, {'from': accounts[0].address})
+    hero.setAddon(spells.address, True, {'from': accounts[0].address})
+    hero.setAddon(loot.address, True, {'from': accounts[0].address})
+
+    return hero
+
+@pytest.fixture(scope="session")
+def HEROS(CC, PP, MP, BookTradable, Hero, spells, loot, items, example_bookmark):
+    ##address _cCA, address _cultureCoin, address _nbt, address _registryAddress
+    hero = accounts[0].deploy(Hero, accounts[0].address, CC.address, example_bookmark.address, MP.address, spells.address, items.address)
+    LOG.info("The heros are here: " + str(hero))
+
+    CC.setAddon(hero.address, True, {'from': accounts[0].address})
+    hero.setAddon(spells.address, True, {'from': accounts[0].address})
+    hero.setAddon(loot.address, True, {'from': accounts[0].address})
+    hero.setAddon(items.address, True, {'from': accounts[0].address})
+    items.setAddon(hero.address, True, {'from': accounts[0].address})
+
+    return hero
+
+@pytest.fixture(scope="session")
+def items(MyItems, CC, MP, spells, example_bookmark):
+
+    ## address _cCA, address _cultureCoin, address _registryAddress, address _baseSpells, address _nbt
+    myItems = accounts[0].deploy(MyItems, accounts[0].address, CC.address, MP.address, spells.address, example_bookmark)
+    spells.setAddon(myItems.address, True, {'from': accounts[0].address})
+    return myItems
+
+@pytest.fixture(scope="session")
+def spells(BaseSpells, proxyAdmin, TransparentUpgradeableProxy, Contract, CC):
+
+    BSImpl = accounts[0].deploy(BaseSpells)
+
+    encoded_initializer_function = encode_function_data(BSImpl.initialize, accounts[0].address, CC.address , "Uri")
+
+    proxy = accounts[0].deploy(TransparentUpgradeableProxy,
+        BSImpl.address,
+        proxyAdmin,
+        encoded_initializer_function
+    )
+
+    bsProxy = Contract.from_abi("BaseSpells", proxy.address, BaseSpells.abi)
+    LOG.info("baseSpells: " + str(bsProxy.address))
+
+    CC.setAddon(bsProxy.address, True, {'from': accounts[0].address})
+
+    return bsProxy
+
+@pytest.fixture(scope="session")
+def loot(CC, BaseLoot, proxyAdmin, TransparentUpgradeableProxy, Contract, spells):   #address _cCA, address _cultureCoin, address _baseSpells
+    lootImpl = accounts[0].deploy(BaseLoot)
+    ## address _cCA, address _cultureCoin, address _baseSpells, string memory _uri
+    encoded_initializer_function = encode_function_data(lootImpl.initialize, accounts[0].address, CC.address, spells.address, "Uri")
+    proxy = accounts[0].deploy(TransparentUpgradeableProxy,
+        lootImpl.address,
+        proxyAdmin,
+        encoded_initializer_function
+    )
+    lootProxy = Contract.from_abi("BaseLoot", proxy.address, BaseLoot.abi)
+    spells.setAddon(lootProxy.address, True, {'from': accounts[0].address})
+
+    return lootProxy
+
+ARCANE_ORB = 1
+
+def test_arcanist(HEROS, spells, arcanist, daedalus):
+    LOG.info("testing the arcanist: " + str(arcanist));
+
+    #res = spells.castRES(HEROS.address, daedalus, arcanist,  {'from': accounts[0].address});
+    #res.info()
+    #print(str(res));
+
+    orbs = spells.balanceOf(HEROS.address, ARCANE_ORB)
+    LOG.info("orbs: " + str(orbs))
+    spells.castAO(HEROS.address, arcanist, accounts[0].address, 3, {'from': accounts[0].address})
+    orbsAfter = spells.balanceOf(HEROS.address, ARCANE_ORB)
+    LOG.info("orbs after: " + str(orbsAfter))
+
+
+@pytest.fixture(scope="session")
+def arcanist(CC, PP, HEROS, spells):
+    LOG.info("test_heros")
+    assert HEROS.address != 0x0
+    assert spells.address != 0x0
+
+    nbt = HEROS.getNBT();
+    tx = PP.buyBook(nbt, {'from': accounts[0].address, 'value': OneCC})
+    #LOG.info("Buy tx: " + str(tx))
+    #tx.info()
+    #LOG.info("Buy tx: " + str(tx.events["PairPrinted"]))
+
+    xBal = CC.balance();
+    LOG.info("X balance: " + str(xBal))
+
+    bal = CC.balanceOf(accounts[0].address)
+    LOG.info("CC Balance of _ : " + str(bal))
+
+    # 7 is probably the right token ... Arcanist = 15
+    myHero = HEROS.heroMint(1, accounts[0].address, 15, 115, {'from': accounts[0].address})
+    LOG.info("My hero: " + str(myHero.return_value))
+    #LOG.info(dir(myHero))
+
+    balNew = CC.balanceOf(accounts[0].address)
+    LOG.info("New CC Balance of _ : " + str(balNew))
+
+    return myHero.return_value;     # This is the magic number of the arcanist
+
+@pytest.fixture(scope="session")
+def daedalus(CC, PP, HEROS, spells):
+    nbt = HEROS.getNBT();
+    daed = HEROS.heroMint(1, accounts[0].address, 134, 100, {'from': accounts[0].address})
+    LOG.info("daealus: " + str(daed.return_value))
+    return daed.return_value;
+
+
+@pytest.fixture(scope="session")
+def warrior(PP, HEROS, spells, arcanist):       # Depends on arcanist...
+    nbt = HEROS.getNBT();
+    myWar = HEROS.heroMint(1, accounts[0].address, 75, 100, {'from': accounts[0].address})
+    LOG.info("My warrior: " + str(myWar.return_value))
+    return myWar.return_value;
+
+def test_warrior(HEROS, warrior, spells, loot, arcanist):
+    LOG.info("testing the warrior: " + str(warrior));
+
+    castsOfAOBefore = HEROS.castsOf(warrior, ARCANE_ORB)
+    LOG.info("Casts of AO before: " + str(castsOfAOBefore))
+    spells.castAO(HEROS.address, arcanist, warrior, 3, {'from': accounts[0].address})
+    castsOfAOAfter = HEROS.castsOf(warrior, ARCANE_ORB)
+    LOG.info("Casts of AO after: " + str(castsOfAOAfter))
+
+    ## address _hero, uint256 _hId, uint256 _target, int _how, uint _amount
+    loot.activateAO(HEROS.address, warrior, arcanist, ARCANE_ORB, 1, {'from': accounts[0].address})
+
+    castsOfAOAfterUsing = HEROS.castsOf(warrior, ARCANE_ORB)
+    LOG.info("Casts of AO after using: " + str(castsOfAOAfterUsing))
+
+
+@pytest.fixture(scope="session")
 def proxyAdmin(ProxyAdmin):
     return accounts[0].deploy(ProxyAdmin)
 
@@ -21,7 +192,7 @@ def CC(CultureCoin, proxyAdmin, TransparentUpgradeableProxy, Contract):
     LOG.info("Deploying CC with amount:" + str(deployAmount))
     CCImpl = accounts[0].deploy(CultureCoin);
 
-    LOG.info(CCImpl.balanceOf(accounts[0].address))
+    #LOG.info(CCImpl.balanceOf(accounts[0].address)) ## Can't be other than 0 now that its upgradable...
 
     encoded_initializer_function = encode_function_data(CCImpl.initialize, deployAmount, accounts[0].address)
 
@@ -33,6 +204,8 @@ def CC(CultureCoin, proxyAdmin, TransparentUpgradeableProxy, Contract):
 
     ccProxy = Contract.from_abi("CultureCoin", proxy.address, CultureCoin.abi)
     LOG.info("CCProxy: " + str(ccProxy.address))
+
+    LOG.info("Your CC balance is: " + str(ccProxy.balanceOf(accounts[0].address)))
 
     return ccProxy # This looks like CC but is the proxy for it.
 
@@ -72,6 +245,7 @@ def scratches(PP, MP):
 
     return myScratches
 
+@pytest.fixture(scope="session")
 def example_bookmark(PP, MP, BookTradable):
     tx = PP.newBookContract("exampleNBT", "BTtestNBT", MP.address, "urlgoeshere", True, maxint, OneCC, 1, accounts[0].address)
     exampleBookMark = BookTradable.at(tx.events[2]["what"])
@@ -114,7 +288,7 @@ def example_book(PP, MP, BookTradable):
 
     return exampleBook
 
-def test_nbts(PP, MP, BookTradable):
+def _test_nbts(PP, MP, BookTradable):
     LOG.info("test_nbts")
 
     example = example_bookmark(PP, MP, BookTradable)
@@ -188,7 +362,7 @@ def gamblers_union(GU, scratches, ben):
     LOG.info("Contest ID: " + str(contestId))
     return contestId
 
-def test_staking_protocol(PP, MP, BookTradable, CC): #, GU, cc_initial_total, cc_initial_balance, gamblers_union):
+def _test_staking_protocol(PP, MP, BookTradable, CC): #, GU, cc_initial_total, cc_initial_balance, gamblers_union):
     LOG.info("test_staking_protocol")
 
     ##CC.setRewardPerHour(1000, {'from': accounts[0].address})
@@ -217,7 +391,7 @@ def pet(ben, scratches):
     assert ben.pet(scratches, "good boy, BEN", 1, {'from': accounts[0].address, 'value': 10000001});
 
 
-def test_new_book_contructor(CC, MP, BookTradable):
+def _test_new_book_contructor(CC, MP, BookTradable):
     #string memory _name, string memory _symbol, address _bookRegistryAddress, string memory _baseuri,
     #                                    bool _burnable, uint256 _maxmint, uint256 _defaultprice, uint256 _defaultfrom, address _gasToken, address _cCA
     bookTradable = accounts[0].deploy(BookTradable, "_name", "_symbol", MP.address, "_baseuri", True, 1, 1, 1, CC.address, accounts[0].address)
@@ -390,7 +564,7 @@ def getCurXMTSPRate(CC, cc_initial_balance, ccTotalSupplyStart):
 
     return ratioXMTSPPerCC
 
-def test_set_dex_rates(CC, cc_warm_dex, ccTotalSupplyStart, cc_initial_balance):
+def _test_set_dex_rates(CC, cc_warm_dex, ccTotalSupplyStart, cc_initial_balance):
     curXRate = CC.getDexXMTSPRate();
     curCCRate = CC.getDexCCRate();
     LOG.info("Current X rate: " + str(curXRate))
@@ -398,7 +572,7 @@ def test_set_dex_rates(CC, cc_warm_dex, ccTotalSupplyStart, cc_initial_balance):
 
     CC.setDexRates(curXRate, curCCRate, {'from': accounts[0].address})
 
-def test_all(gamblers_union): #cc_dex, gamblers_union):
+def test_all(): #gamblers_union): #cc_dex, gamblers_union):
     pass
 
 @pytest.fixture(scope="session")
